Temporarily switch data type to 32-bit float

+++ scos_actions/actions/acquire_sea_data_product.py
@@ -287,7 +287,7 @@ class NasctnSeaDataProduct(Action):
         # Reduce data types to half-precision floats
         tic = perf_counter()
         for i in range(len(data_product)):
-            data_product[i] = data_product[i].astype(np.half)
+            data_product[i] = data_product[i].astype(np.single)
         toc = perf_counter()
         logger.debug(f"Reduced data types to half-precision float in {toc-tic:.2f} s")
